Remove legacy flashblocks RPC tests
